Voldemort serialization should work when requiring 'rational'
The voldemort client cannot serialize nil and 0.0 if 'rational' is
required.  This is because DOUBLE_MIN_VAL, defined as "2 ** -1074",
becomes a Rational type when 'rational' is required.  If you try
comparing this to 0.0 it returns true because Ruby fails to cast the
Rational type to a float (because it is so large).  Likewise,
it can't serialize nil either because the Rational value cannot be
packed (again, because it can't convert it to a float).

This behavior occurs because 'rational' overrides Fixnum.  The fix is
simple.  Just use "2.0 ** -1074" instead.

I discovered this trying to use voldemort-rb within cucumber.  Cucumber
requires 'yaml', which requires 'date', etc. and eventually 'rational'.